+++ src/Ampersand/Misc/Options.hs
@@ -40,8 +40,10 @@ data Options = Options { environment :: EnvironmentOptions
                        , zwolleVersion :: String -- the version in github of the prototypeFramework. can be a tagname, a branchname or a SHA
                        , forceReinstallFramework :: Bool -- when true, an existing prototype directory will be destroyed and re-installed
                        , dirCustomizations :: [FilePath] -- the directory that is copied after generating the prototype
+                       , runComposer :: Bool -- if True, runs Composer (php package manager) when generating prototype. Requires PHP and Composer on the machine. Added as switch to disable when building with Docker.
                        , allInterfaces :: Bool
                        , runAsDaemon :: Bool -- run Ampersand as a daemon. (for use with the vscode extension)
+                       , daemonConfig :: FilePath -- the path (relative from current directory OR absolute) and filename of a file that contains the root file(s) to be watched by the daemon.
                        , dbName :: String
                        , namespace :: String
                        , testRule :: Maybe String
@@ -210,9 +212,10 @@ getOptions' envOpts =
                       , outputfile       = fatal "No monadic options available."
                       , dirPrototype     = fromMaybe "." (envDirPrototype envOpts) </> (takeBaseName (fromMaybe "" fName)) <.> ".proto"
                       , dirSource        = takeDirectory $ fromMaybe "/" fName
-                      , zwolleVersion    = "v1.1.1"
+                      , zwolleVersion    = "v1.2.0"
                       , forceReinstallFramework = False
                       , dirCustomizations = ["customizations"]
+                      , runComposer      = True -- by default run Composer (php package manager) when deploying prototype for backward compatibility
                       , dbName           = fmap toLower . fromMaybe ("ampersand_" ++ takeBaseName (fromMaybe "prototype" fName)) $ envDbName envOpts
                       , dirExec          = takeDirectory (envExePath envOpts)
                       , preVersion       = fromMaybe "" $ envPreVersion envOpts
@@ -226,6 +229,7 @@ getOptions' envOpts =
                       , genPrototype     = False
                       , allInterfaces    = False
                       , runAsDaemon      = False
+                      , daemonConfig     = ".ampersand"
                       , namespace        = ""
                       , testRule         = Nothing
               --        , customCssFile    = Nothing
@@ -359,7 +363,7 @@ options = [ (Option ['v']   ["version"]
             , Public)
           , (Option ['h','?'] ["help"]
                (NoArg (\opts -> opts{showHelp = True}))
-               "get (this) usage information."
+               "get (this) usage information. Add --verbose for more advanced options."
             , Public)
           , (Option ['V']   ["verbose"]
                (NoArg (\opts -> opts{ verboseP  = True
@@ -398,42 +402,46 @@ options = [ (Option ['v']   ["version"]
                (ReqArg (\x opts -> opts {zwolleVersion = x}
                        ) "VERSION")
                ("tag, branch or SHA of the prototype framework on Github. (What purpose does this serve?)")
-            , Public)
+            , Hidden)
           , (Option []      ["force-reinstall-framework"]
                (NoArg (\opts -> opts{forceReinstallFramework = True}))
                "re-install the prototype framework. This discards any previously installed version."
-            , Public)
+            , Hidden)
           , (Option []     ["customizations"]
                (ReqArg (\names opts -> opts {dirCustomizations = splitOn ";" names}
                        ) "DIRECTORY")
                "copy a directory into the generated prototype, overriding the default directory called 'customizations'."
-            , Public)
+            , Hidden)
+          , (Option []      ["skip-composer"]
+               (NoArg (\opts -> opts{runComposer = False}))
+               "skip installing php dependencies (using Composer) for prototype framework."
+            , Hidden)
           , (Option ['d']  ["dbName"]
                (ReqArg (\nm opts -> opts{dbName = if nm == ""
                                                          then dbName opts
                                                          else map toLower nm}
                        ) "NAME")
                ("database name (This overrules environment variable "++ dbNameVarName ++ ", defaults to filename) to which the prototype will connect for persistent storage.")
-            , Public)
+            , Hidden)
           , (Option []  ["sqlHost"]
                (ReqArg (\nm opts -> opts{sqlHost = if nm == ""
                                                           then sqlHost opts
                                                           else nm}
                        ) "HOSTNAME")
                "set SQL host name (Defaults to `localhost`), to identify the host on which the persistent store resides"
-            , Public)
+            , Hidden)
           , (Option []  ["sqlLogin"]
                (ReqArg (\nm opts -> opts{sqlLogin = if nm == ""
                                                           then sqlLogin opts
                                                           else nm}
                        ) "USER")
                "set SQL user name (Defaults to `ampersand`), to let your application login to the database."
-            , Public)
+            , Hidden)
           , (Option []  ["sqlPwd"]
                (ReqArg (\nm opts -> opts{sqlPwd = nm}
                        ) "PASSWORD")
                "set SQL password (Defaults to `ampersand`), to let your application login to the database."
-            , Public)
+            , Hidden)
           , (Option []        ["sql-bin-tables"]
                (NoArg (\opts -> opts{sqlBinTables = True}))
                "generate binary tables only in SQL database, for testing purposes."
@@ -441,14 +449,16 @@ options = [ (Option ['v']   ["version"]
           , (Option ['x']     ["interfaces"]
                (NoArg (\opts -> opts{allInterfaces  = True}))
                "generate interfaces, which currently does not work."
-            , Public)
-          , (Option []        ["daemon"]
-               (NoArg (\opts -> opts{runAsDaemon = True}))
-               "Run ampersand as daemon, for use by the vscode ampersand-language-extention."
             , Hidden)
+          , (Option []        ["daemon"]
+               (OptArg (\fn opts -> opts{runAsDaemon = True
+                                        ,daemonConfig = fromMaybe (daemonConfig opts) fn
+                                        })"configfile")
+               "Run ampersand as daemon, for use by the vscode ampersand-language-extention. An optional parameter may be specified to tell what config file is used. This defaults to `.ampersand`."
+            , Public)
           , (Option ['e']     ["export"]
                (OptArg (\mbnm opts -> opts{export2adl = True
-                                                   ,outputfile = fromMaybe "Export.adl" mbnm}) "file")
+                                          ,outputfile = fromMaybe "Export.adl" mbnm}) "file")
                "export as plain Ampersand script, for round-trip testing of the Ampersand compiler."
             , Public)
           , (Option ['o']     ["outputDir"]
@@ -460,7 +470,7 @@ options = [ (Option ['v']   ["version"]
                (ReqArg (\nm opts -> opts{namespace = nm}
                        ) "NAMESPACE")
                "prefix database identifiers with this namespace, to isolate namespaces within the same database."
-            , Public)
+            , Hidden)
           , (Option ['f']   ["fspec"]
                (ReqArg (\w opts -> opts
                                 { genFSpec=True
@@ -511,19 +521,19 @@ options = [ (Option ['v']   ["version"]
           , (Option []        ["sqldump"]
                (NoArg (\opts -> opts{sqlDump = True}))
                "generate a dump of SQL queries (for debugging)."
-            , Public)
+            , Hidden)
           , (Option []        ["crowfoot"]
                (NoArg (\opts -> opts{crowfoot = True}))
                "generate crowfoot notation in graphics, to please crowfoot addicts."
-            , Public)
+            , Hidden)
           , (Option []        ["blackWhite"]
                (NoArg (\opts -> opts{blackWhite = True}))
                "avoid coloring conventions to facilitate readable pictures in black and white."
-            , Public)
+            , Hidden)
           , (Option []        ["altGraphics"]
                (NoArg (\opts -> opts{doubleEdges = not (doubleEdges opts)}))
                "generate graphics in an alternate way. (you may experiment with this option to see the differences for yourself)"
-            , Public)
+            , Hidden)
           , (Option []        ["noGraphics"]
                (NoArg (\opts -> opts{noGraphics = True}))
                "omit the generation of graphics during generation of the functional design document to speed up the compiler."
@@ -555,7 +565,7 @@ options = [ (Option ['v']   ["version"]
           , (Option []        ["pop-xlsx"]
                (NoArg (\opts -> opts{genPOPExcel = True}))
                "Generate an .xmlx file containing the populations of your script."
-            , Public)
+            , Hidden)
           , (Option []        ["language"]
                (ReqArg (\l opts-> opts{language = case map toUpper l of
                                                        "NL"  -> Just Dutch
@@ -609,7 +619,7 @@ options = [ (Option ['v']   ["version"]
           , (Option []        ["do-not-trim-cellvalues"]
                (NoArg (\opts -> opts{trimXLSXCells = False}))
                "Do not ignore leading and trailing spaces in .xlsx files that are INCLUDED in the script." -- :-)
-            , Public)
+            , Hidden)
           ]
 
 usageInfo' :: Options -> String