+++ src/Ampersand/Output/ToJSON/Populations.hs
@@ -48,7 +48,7 @@ instance JSON (MultiFSpecs,Bool) Populations where
      where ftl = fatal "There is no grinded fSpec."
 instance JSON (A_Concept,Bool) AtomValuesOfConcept where
  fromAmpersand _ multi (cpt,doMeta) = AtomValuesOfConcept
-   { avcJSONconcept = Text.pack (escapeIdentifier . name $ cpt)
+   { avcJSONconcept = Text.pack (idWithoutType cpt)
    , avcJSONatoms   = map (Text.pack . showValADL) (Set.elems $ atomsBySmallestConcept theFSpec cpt)
    }
   where